gosym: split back to upstream files
Split gosym.go to pclntab.go and symtab.go to match the upstream layout
to make it simpler to merge upstream changes.
